requirements.txt
# Servidor Web
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# RAG e IA
langchain==0.3.0
langchain-community==0.3.0
chromadb==0.4.24
anthropic==0.39.0
sentence-transformers==3.2.0

# Processamento de Documentos
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2

# Autenticação e Segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilitários
tiktoken==0.8.0
pydantic==2.9.0
aiofiles==24.1.0
requests==2.31.0
